+++ flake8_tidy_imports.py
@@ -7,7 +7,7 @@ import flake8
 
 __author__ = 'Adam Johnson'
 __email__ = 'me@adamj.eu'
-__version__ = '1.0.6'
+__version__ = '1.1.0'
 
 
 class ImportChecker(object):